+++ hclasses/CTString.cpp
@@ -139,7 +139,6 @@ namespace CT {
       strings[j].count = n;
       delete token;
     }
-    CTlink(strings, n);
     return strings;
   };
 
@@ -666,7 +665,7 @@ namespace CT {
   }
 
   float string::toFloat() {
-    float fValue = (float)atof(c_str());
+    float fValue = (float)atof(trim().c_str());
     return fValue;
   }
 
@@ -756,24 +755,24 @@ namespace CT {
   }
 
   bool string::isFloat() {
-    const size_t inputLength = this->length();
-    const char *inputStr = this->c_str();
+    CT::string inputStr = this->trim().c_str();
+
     /*check size */
-    if (this->empty() || inputLength > CT_MAX_NUM_CHARACTERS_FOR_FLOAT) {
+    if (inputStr.empty() || inputStr.length() > CT_MAX_NUM_CHARACTERS_FOR_FLOAT) {
       return false;
     }
     /* NaN is a "float"...in this context */
-    if (this->equals("NaN")) {
+    if (inputStr.equals("NaN")) {
       return true;
     }
 
-    if (std::regex_match(inputStr, isFloatRegex)) {
+    if (std::regex_match(inputStr.c_str(), isFloatRegex)) {
       return true;
     }
     return false;
   }
 
-  string getHex(unsigned int number) {
+  string string::getHex(unsigned int number) {
     int hex = number % 256;
     unsigned char a = hex / 16;
     unsigned char b = hex % 16;